Move new tests to dedicated file.
